updates by Suchandradi

+++ interface/Reconstruction.h
@@ -10,11 +10,11 @@ class TH1;
 using Utility::Cluster;
 class Reconstruction : public tBeamBase {
  public:
-  Reconstruction(const string inFilename,const string outFilename,int stubWindow, bool publish);
+  Reconstruction(const string inFilename,const string outFilename,float stubWindow, bool publish);
   ~Reconstruction(){}
   void bookHistogram(TFile* fout);
   void publishPlots(TString dirName);
-  void setStubWindow(int stubWindow) { stubWindow_ = stubWindow; }
+  void setStubWindow(float stubWindow) { stubWindow_ = stubWindow; }
   void beginJob();
   void Loop();
   void endJob();
@@ -22,7 +22,7 @@ class Reconstruction : public tBeamBase {
  private:
   TString outFile;
   TFile* fout_;
-  int stubWindow_;
+  float stubWindow_;
   bool publishPng_;
   std::map<std::string,std::vector<Cluster> >  detClustermap_;
 };
