+++ interface/ReconstructionFromRaw.h
@@ -10,11 +10,11 @@ class TH1;
 using Utility::Cluster;
 class ReconstructionFromRaw : public trawTupleBase {
  public:
-  ReconstructionFromRaw(const string inFilename,const string outFilename,int stubWindow, bool publish);
+  ReconstructionFromRaw(const string inFilename,const string outFilename,float stubWindow, bool publish);
   ~ReconstructionFromRaw(){}
   void publishPlots(TString dirName);
   void bookHistogram(TFile* fout);
-  void setStubWindow(int stubWindow) { stubWindow_ = stubWindow; }
+  void setStubWindow(float stubWindow) { stubWindow_ = stubWindow; }
   void beginJob();
   void Loop();
   void endJob();
@@ -22,7 +22,7 @@ class ReconstructionFromRaw : public trawTupleBase {
  private:
   TString outFile;
   TFile* fout_;
-  int stubWindow_;
+  float stubWindow_;
   bool publishPng_;
   std::map<std::string,std::vector<Cluster> >  detClustermap_;
 };
